Increase frame count size to 2 bytes

+++ inc/actors.h
@@ -37,7 +37,7 @@ typedef struct SpwData{
 typedef struct Actor{
     const Entity * character;
     u8 status;
-    u8 frames;
+    u16 frames;
     u16 pos[2]; //px when declared, px/DECIMALS when in use.
     s8 speed[2];
     u16 timer;
